+++ local/bin/shell_go.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+docker-compose exec -u $UID golang bash